requirements.txt
fastapi 
uvicorn 
sqlalchemy 
psycopg2-binary 
google-auth 
google-auth-oauthlib 
google-auth-httplib2 
google-api-python-client 
redis